Move StateSnapshot to its own file
